deal_service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.10
psycopg2-binary==2.9.5
python-jose==3.3.0
pandas==1.5.3